% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsNtables.R
\name{backtestBoxPlot}
\alias{backtestBoxPlot}
\title{Plots a boxplot from a backtest}
\usage{
backtestBoxPlot(backtest, measure = "Annual volatility",
  type = c("ggplot2", "simple"), mar = c(3, 10, 3, 1), ...)
}
\description{
Plots a boxplot from a backtest
}
\examples{
# After running a backtest and extracting a performance summary, one would get:
bt <- portfolioBacktest(...)

# Here let's generate fake random data for simplicity  
TODO (Rui): needs to be fixed

fun1_1 <- list()
fun1_1$performance <- runif(1)
names(fun1_1$performance) <- "Sharpe ratio"
fun1_2 <- list()
fun1_2$performance <- runif(1)
names(fun1_2$performance) <- "Sharpe ratio"
fun1 <- list(fun1_1, fun1_1)
fun2 <- list(fun1_1, fun1_1)
bt <- list("fun1" = fun1, "fun2" = fun2)

# Now we can plot
backtestBoxPlot(bt, "Sharpe ratio")

}
\author{
Daniel P. Palomar and Rui Zhou
}
