% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtestPortfolio-package.R
\docType{package}
\name{backtestPortfolio-package}
\alias{backtestPortfolio-package}
\title{backtestPortfolio: XXXX}
\description{
Backtest of a portfolio design with real data taken randomly from different
markets and selecting randomly a subset of stocks for proper assessments.
It also allows to backtest multiple portfolio designs and rank them based
on a number of criteria including expected return, Sharpe ratio, drawdown,
etc.
}
\section{Functions}{

\code{\link{spIndexTrack}}
}

\section{Help}{

For a quick help see the README file:
\href{https://rawgit.com/dppalomar/sparseIndexTracking/master/README.html}{GitHub-README} and
\href{https://cran.r-project.org/web/packages/sparseIndexTracking/README.html}{CRAN-README}.

For more details see the vignette:
\href{https://rawgit.com/dppalomar/sparseIndexTracking/master/vignettes/SparseIndexTracking-vignette.html}{GitHub-html-vignette},
\href{https://rawgit.com/dppalomar/sparseIndexTracking/master/vignettes/SparseIndexTracking-vignette.pdf}{GitHub-pdf-vignette}, and
\href{https://cran.r-project.org/web/packages/sparseIndexTracking/vignettes/SparseIndexTracking-vignette.pdf}{CRAN-pdf-vignette}.
}

\references{
K. Benidis, Y. Feng, and D. P. Palomar, "Sparse Portfolios for High-Dimensional Financial Index Tracking,"
\emph{IEEE Transactions on Signal Processing}, vol. 66, no. 1, pp. 155-170, Jan. 2018.
}
\author{
Konstantinos Benidis and Daniel P. Palomar
}
