% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtestSummary.R
\name{backtestSelector}
\alias{backtestSelector}
\title{Portfolio Backtest Results Selector}
\usage{
backtestSelector(res, portfolio_name = NA, portfolio_index = NA,
  selector = NULL)
}
\arguments{
\item{res}{the results from function `portfolioBacktest()`}

\item{portfolio_name}{the name of a portfolio}

\item{portfolio_index}{the index of a portfolio}

\item{selector}{a vector of required performance}
}
\value{
a list of desired results
}
\description{
Select the results from portfolio backtest and return as a matrix
}
\author{
Daniel P. Palomar and Rui Zhou
}
