% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplePortfolioBacktest.R
\name{checkUninstalledPackages}
\alias{checkUninstalledPackages}
\title{Checking uninstalled packages written in the portfolio functions defined by customer}
\usage{
checkUninstalledPackages(folder_path)
}
\arguments{
\item{folder_path}{Path for a folder which contains all (and only) functions to be evaluated}
}
\description{
Checke uninstalled packages of portfolio functions written in format form
}
\author{
Daniel P. Palomar and Rui Zhou
}
