% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Random sets of stock market data from the S&P 500, over random periods of two years with a random selection of 50 stocks of each universe.}
\format{A list of \code{xts} objects with the prices of random sets of stock market data from the S&P 500 over random periods of two years with a random selection of 50 stocks of each universe.}
\source{
\href{https://finance.yahoo.com}{Yahoo! Finance}
}
\usage{
data(dataset)
}
\description{
Random sets of stock market data from the S&P 500, over random periods of two years with a random selection of 50 stocks of each universe.
}
\keyword{datasets}
