% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_tuning.R
\name{genRandomFuns}
\alias{genRandomFuns}
\title{Generate multiple versions of a function with randomly chosen parameters}
\usage{
genRandomFuns(portfolio_fun, params_grid, name = "portfolio",
  N_realizations = NULL)
}
\arguments{
\item{portfolio_fun}{Portfolio function with parameters unspecified.}

\item{params_grid}{Named list containing for each parameter the possible values it can take.}

\item{name}{String with the name of the portfolio function.}

\item{N_realizations}{Number of random realizations of the parameters.}
}
\description{
Portfolio functions usually contain some parameters that can be tuned. 
This function creates multiple versions of a function with randomly chosen parameters.
}
\examples{
# define GMVP with parameters "lookback" and "N_stocks"
GMVP_portfolio_fun <- function(data) {
  prices <- tail(dataset$adjusted, lookback)
  prices <- prices[, 1:N_stocks]
  X <- diff(log(prices))[-1]
  Sigma <- cov(X)
  # design GMVP
  w <- solve(Sigma, rep(1, nrow(Sigma)))
  return(w/sum(w))
}

portfolio_list <- genRandomFuns(portfolio_fun = GMVP_portfolio_fun, 
                                params_grid = list(lookback = c(100, 120, 140, 160),
                                                    N_stocks = c(5, 10, 20, 30, 40)),
                                name = "GMVP", 
                                N_realizations = 10)
names(portfolio_list)
portfolio_list[[1]]
rlang::env_print(portfolio_list[[1]])
rlang::fn_env(portfolio_list[[1]])$lookback
rlang::fn_env(portfolio_list[[1]])$N_stocks

}
\author{
Daniel P. Palomar and Rui Zhou
}
