% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_portfolios.R
\name{listPortfoliosWithFailures}
\alias{listPortfoliosWithFailures}
\title{List portfolios with failures}
\usage{
listPortfoliosWithFailures(bt)
}
\arguments{
\item{bt}{Backtest results as produced by the function \code{\link{portfolioBacktest}}.}
}
\description{
List portfolios with failures
}
\examples{
\donttest{
library(portfolioBacktest)
data(dataset10)  # load dataset

# define your own portfolio function
portfolio_with_errors <- function(dataset, ...) {
  return(NA)
}

# do backtest
bt <- portfolioBacktest(list("Portfolio with errors" = portfolio_with_errors), dataset10)

listPortfoliosWithFailures(bt)
}

}
\author{
Daniel P. Palomar
}
