% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtestPortfolio.R
\name{multipleBacktestPortfolio}
\alias{multipleBacktestPortfolio}
\title{Backtesting Portfolio Design on a Rolling-Window Basis of Set of Prices}
\usage{
multipleBacktestPortfolio(portfolio_fun, prices, shortselling = FALSE,
  leverage = 1, T_sliding_window = 6 * 21, freq_optim = 5,
  freq_rebalance = freq_optim)
}
\arguments{
\item{portfolio_fun}{function that takes as input an \code{xts} containing the stock prices and returns the portfolio weights.}

\item{prices}{a list of \code{xts} containing the stock prices for the backtesting.}

\item{shortselling}{whether shortselling is allowed or not (default \code{FALSE}).}

\item{leverage}{amount of leverage (default is 1, so no leverage).}

\item{T_sliding_window}{length of the sliding window.}

\item{freq_optim}{how often the portfolio is to be reoptimized.}

\item{freq_rebalance}{how often the portfolio is to be rebalanded.}
}
\value{
A list containing the performance in the following elements:
\item{\code{TBD}  }{m-by-m matrix, columns corresponding to eigenvectors.}
\item{\code{TBD}  }{m-by-1 vector corresponding to eigenvalues.}
}
\description{
Backtest a portfolio design contained in a function on a rolling-window basis of a set of prices.
}
\examples{
library(backtestPortfolio)
library(xts)

# load data
data(prices)

# define portfolio function
portfolio_fun <- function(prices) {
  X <- diff(log(prices))[-1]  # compute log returns
  Sigma <- cov(X)  # compute SCM
  # design GMVP
  w <- solve(Sigma, rep(1, nrow(Sigma)))
  w <- w/sum(w)
  return(w)
}

# perform backtesting
res <- backtestPortfolio(portfolio_fun, prices[1:5])
print(res)

}
\author{
Daniel P. Palomar and Rui Zhou
}
