% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplePortfolioBacktest.R
\name{multiplePortfolioBacktest}
\alias{multiplePortfolioBacktest}
\title{Evaluating multiple portfolio functions defined by customer}
\usage{
multiplePortfolioBacktest(folder_path = NULL,
  portfolio_fun_list = NULL, prices__, return_all = FALSE, ...)
}
\arguments{
\item{folder_path}{path for a folder which contains all (and only) functions to be evaluated}

\item{portfolio_fun_list}{a list of portfolio functions, valid when \code{folder_path} is not passed}

\item{return_all}{logical, indicating whether return all the results.}

\item{prices}{a list of \code{xts} containing the stock prices for the backtesting.}

\item{return_portfolio}{logical value, whether return portfolios.}

\item{shortselling}{whether shortselling is allowed or not (default \code{FALSE}).}

\item{leverage}{amount of leverage (default is 1, so no leverage).}

\item{T_rolling_window}{length of the rolling window.}

\item{optimize_every}{how often the portfolio is to be optimized.}

\item{rebalance_every}{how often the portfolio is to be rebalanced.}

\item{time_limit}{time limit for executing portfolio function on a single data set}
}
\value{
A list containing the performance in the following elements:
\item{\code{stud_names}  }{string vector, recording the student names extracted from files' name.}
\item{\code{stud_IDs}  }{string vector, recording the student IDs extracted from files' name.}
\item{\code{performance_summary}}{matrix, indicating each student's performance summary in each row}
\item{\code{time_average}}{vector, recording the average execution time for successful application of portfolio functions}
\item{\code{failure_ratio}}{vector, recording the failure ratio of applying given portfolio function to different data examples}
\item{\code{error_message}}{string list, recording the error message when error happens}
}
\description{
Evaluate multiple portfolio functions written in format form
}
\author{
Daniel P. Palomar and Rui Zhou
}
