% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplePortfolioFunEval.R
\name{multiplePortfolioFunEval}
\alias{multiplePortfolioFunEval}
\title{Evaluating multiple portfolio functions defined by customer}
\usage{
multiplePortfolioFunEval(path, prices, shortselling = FALSE,
  leverage = 1, T_sliding_window = 6 * 21, freq_optim = 5,
  freq_rebalance = freq_optim)
}
\arguments{
\item{path}{absolute path for a folder which contains all (and only) functions to be evaluated}

\item{prices}{a list of \code{xts} containing the stock prices for the backtesting.}

\item{shortselling}{whether shortselling is allowed or not (default \code{FALSE}).}

\item{leverage}{amount of leverage (default is 1, so no leverage).}

\item{T_sliding_window}{length of the sliding window.}

\item{freq_optim}{how often the portfolio is to be reoptimized.}

\item{freq_rebalance}{how often the portfolio is to be rebalanded.}
}
\value{
A list containing the performance in the following elements:
\item{\code{TBD}  }{m-by-m matrix, columns corresponding to eigenvectors.}
\item{\code{TBD}  }{m-by-1 vector corresponding to eigenvalues.}
}
\description{
Evaluate multiple portfolio functions written in format form
}
\author{
Daniel P. Palomar and Rui Zhou
}
