% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices1.R
\docType{data}
\name{prices1}
\alias{prices1}
\title{Random selection of 30 stocks from the S&P 500}
\format{An \code{xts} object with the prices.}
\source{
\href{https://finance.yahoo.com}{Yahoo! Finance}
}
\usage{
data(prices1)
}
\description{
Random selection of 30 stocks from the S&P 500
}
\keyword{datasets}
