% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices2.R
\docType{data}
\name{prices2}
\alias{prices2}
\title{Random selection of 50 stocks from the S&P 500}
\format{An \code{xts} object with the prices.}
\source{
\href{https://finance.yahoo.com}{Yahoo! Finance}
}
\usage{
data(prices2)
}
\description{
Random selection of 50 stocks from the S&P 500
}
\keyword{datasets}
