% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioBacktest.R
\name{singlePortfolioBacktest}
\alias{singlePortfolioBacktest}
\title{Backtesting on a single xts}
\usage{
singlePortfolioBacktest(portfolio_fun, prices, shortselling = FALSE,
  leverage = 1, T_sliding_window = 6 * 21, freq_optim = 5,
  freq_rebalance = freq_optim)
}
\description{
Backtesting on a single xts
}
