% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP500_symbols.R
\docType{data}
\name{SP500_symbols}
\alias{SP500_symbols}
\title{The stocks symbols from S&P 500 Index list}
\format{A string vector of stock symbols from S&P 500 Index list. 
        The market index symbol is concluded as the attribute "index_symbol".}
\source{
\href{https://finance.yahoo.com}{Yahoo! Finance}
}
\usage{
data(SP500_symbols)
}
\description{
The stocks symbols from S&P 500 Index list
}
\keyword{SP500_symbols}
