% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsNtables.R
\name{summaryBarPlot}
\alias{summaryBarPlot}
\title{Plots a barplot of a backtest summary}
\usage{
summaryBarPlot(res_summary, measures = NULL, type = c("ggplot2",
  "simple"), mar = c(3, 3, 3, 11), inset = c(-0.3, 0),
  legend_loc = "right", ...)
}
\description{
Plots a barplot of a backtest summary
}
\examples{
library(portfolioBacktest)
data("dataset10")  # load dataset

# define your own portfolio function
quintile_portfolio <- function(data) {
  X <- diff(log(data$adjusted))[-1]  
  N <- ncol(X)
  ranking <- sort(colMeans(X), decreasing = TRUE, index.return = TRUE)$ix
  w <- rep(0, N)
  w[ranking[1:round(N/5)]] <- 1/round(N/5)
  return(w)
}

# do backtest
bt <- portfolioBacktest(list("Quintile" = quintile_portfolio), dataset10,
                        benchmark = c("uniform", "index"))
                        
# Now we can obtain the table
summaryBarPlot(res_summary_median, measures = c("max drawdown", "annual volatility"))
summaryBarPlot(res_summary_median, measures = c("max drawdown", "annual volatility"), type = "ggplot2")

}
\author{
Daniel P. Palomar and Rui Zhou
}
