% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsNtables.R
\name{backtestBoxPlot}
\alias{backtestBoxPlot}
\title{Plots a boxplot from a backtest}
\usage{
backtestBoxPlot(backtest, measure = "Annual volatility",
  type = c("ggplot2", "simple"), mar = c(3, 10, 3, 1), ...)
}
\description{
Plots a boxplot from a backtest
}
\examples{
library(portfolioBacktest)
data("dataset10")  # load dataset

# define your own portfolio function
quintile_portfolio <- function(data) {
  X <- diff(log(data$adjusted))[-1]  
  N <- ncol(X)
  ranking <- sort(colMeans(X), decreasing = TRUE, index.return = TRUE)$ix
  w <- rep(0, N)
  w[ranking[1:round(N/5)]] <- 1/round(N/5)
  return(w)
}

# do backtest
bt <- portfolioBacktest(list("Quintile" = quintile_portfolio), dataset10,
                        benchmark = c("uniform", "index"))

# Now we can plot
backtestBoxPlot(bt, "Sharpe ratio")

}
\author{
Daniel P. Palomar and Rui Zhou
}
