% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtestSummary.R
\name{backtestSummary}
\alias{backtestSummary}
\title{Portfolio Backtest Results Summary}
\usage{
backtestSummary(res, portfolio_names = NA, portfolio_indexs = NA,
  summary_fun = median, show_benchmark = TRUE)
}
\arguments{
\item{res}{the results from function `portfolioBacktest()`}

\item{portfolio_names}{the names of a portfolio}

\item{portfolio_indexs}{the indexs of a portfolio}

\item{show_benchmark}{logical value indicating whether to show benchmark in the portfolio}

\item{summary_funs}{summary function}
}
\value{
a list of desired results
}
\description{
Summarize the results from portfolio backtest
}
\examples{
library(portfolioBacktest)
data("dataset10")  # load dataset

# define your own portfolio function
uniform_portfolio <- function(dataset) {
  N <- ncol(dataset$adjusted)
  return(rep(1/N, N))
}

# do backtest
bt <- portfolioBacktest(list("Uniform" = uniform_portfolio), dataset10)

# show the summary
bt_sum <- backtestSummary(bt)
bt_sum$performance_summary

}
\author{
Daniel P. Palomar and Rui Zhou
}
