% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioBacktest-package.R
\docType{package}
\name{portfolioBacktest-package}
\alias{portfolioBacktest-package}
\title{portfolioBacktest: Backtesting of a Portfolio Over Multiple Datasets}
\description{
portfolioBacktest: Backtesting of a Portfolio Over Multiple Datasets
}
\section{Functions}{

\code{\link{portfolioBacktest}}, \code{\link{multiplePortfolioBacktest}}
}

\section{Help}{

For a quick help see the README file:
\href{https://rawgit.com/dppalomar/portfolioBacktest/master/README.html}{GitHub-README} and
\href{https://cran.r-project.org/web/packages/portfolioBacktest/README.html}{CRAN-README}.
}

\author{
Daniel P. Palomar and Rui ZHOU
}
