% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioBacktest-package.R
\docType{package}
\name{portfolioBacktest-package}
\alias{portfolioBacktest-package}
\title{portfolioBacktest: Automatic Backtesting of Portfolios Over Multiple Datasets}
\description{
portfolioBacktest: Automatic Backtesting of Portfolios Over Multiple Datasets
}
\section{Functions}{

\code{\link{portfolioBacktest}}, \code{\link{multiplePortfolioBacktest}}
}

\section{Help}{

For a quick help see the README file:
\href{https://CRAN.R-project.org/package=portfolioBacktest/readme/README.html}{CRAN-README} and
\href{https://raw.githack.com/dppalomar/portfolioBacktest/master/README.html}{GitHub-README}.

For more details see the vignette:
\href{https://CRAN.R-project.org/package=portfolioBacktest/vignettes/PortfolioBacktest.html}{CRAN-html-vignette},
\href{https://CRAN.R-project.org/package=portfolioBacktest/vignettes/PortfolioBacktest-pdf.pdf}{CRAN-pdf-vignette},
\href{https://raw.githack.com/dppalomar/portfolioBacktest/master/vignettes/PortfolioBacktest.html}{GitHub-html-vignette}, and
\href{https://docs.google.com/viewer?url=https://github.com/dppalomar/portfolioBacktest/raw/master/vignettes/PortfolioBacktest-pdf.pdf}{GitHub-pdf-vignette}.
}

\author{
Rui ZHOU and Daniel P. Palomar
}
