% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioLeaderboard.R
\name{portfolioLeaderboard}
\alias{portfolioLeaderboard}
\title{The leaderboard of portfolio function or students}
\usage{
portfolioLeaderboard(res = NA, weights = list())
}
\arguments{
\item{res}{Returned result from function \code{multiplePortfolioBacktest}}

\item{weights}{a list of weights for \code{sharpe ratio}, \code{max drawdown}, \code{expected return}, \code{volatility} 
\code{cpu time} and \code{failure ratio}}
}
\value{
a matrix as the leaderboard
}
\description{
Ranking the portfolio functions according to passed criteria
}
\author{
Daniel P. Palomar and Rui Zhou
}
