% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioLeaderboard.R
\name{portfolioLeaderboard}
\alias{portfolioLeaderboard}
\title{The leaderboard of portfolio function or students}
\usage{
portfolioLeaderboard(res = NA, weights = list(),
  summary_fun = median, show_benchmark = TRUE)
}
\arguments{
\item{res}{Returned result from function \code{multiplePortfolioBacktest}}

\item{weights}{a list of weights for \code{Sharpe ratio}, \code{max drawdown}, \code{expected return}, \code{volatility} 
\code{Sterling ratio}, \code{Omega ratio}, \code{ROT}, \code{cpu time} and \code{failure ratio}}
}
\value{
a matrix as the leaderboard
}
\description{
Ranking the portfolio functions according to passed criteria
}
\examples{
library(portfolioBacktest)
data("dataset10")  # load dataset

# define your own portfolio function
quintile_portfolio <- function(data) {
  X <- diff(log(data$adjusted))[-1]  
  N <- ncol(X)
  ranking <- sort(colMeans(X), decreasing = TRUE, index.return = TRUE)$ix
  w <- rep(0, N)
  w[ranking[1:round(N/5)]] <- 1/round(N/5)
  return(w)
}

# do backtest
bt <- portfolioBacktest(list("Quintile" = quintile_portfolio), dataset10,
                        benchmark = c("uniform", "index"))

# show leaderboard
leaderboard <- portfolioLeaderboard(bt, weights = list("Sharpe ratio"  = 7,
                                                       "max drawdown"  = 1,
                                                       "annual return" = 1,
                                                       "ROT bps"       = 1))
leaderboard$leaderboard_scores

}
\author{
Daniel P. Palomar and Rui Zhou
}
