% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\docType{data}
\name{prices}
\alias{prices}
\title{Set of random selections of 50 stocks from the S&P 500}
\format{A list of \code{xts} objects with the random selection of 50 stock price during random periods of two years.}
\source{
\href{https://finance.yahoo.com}{Yahoo! Finance}
}
\usage{
data(prices)
}
\description{
Set of random selections of 50 stocks from the S&P 500
}
\keyword{datasets}
