% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP500_symbols.R
\docType{data}
\name{SP500_symbols}
\alias{SP500_symbols}
\title{Stock symbols of the S&P 500 constituents}
\format{String vector of stock symbols of the S&P 500 constituents. 
        The market index symbol is concluded as the attribute "index_symbol".}
\source{
\href{https://finance.yahoo.com}{Yahoo! Finance}
}
\usage{
data(SP500_symbols)
}
\description{
Stock symbols of the S&P 500 constituents
}
\keyword{SP500_symbols}
