% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockDataDownload.R
\name{stockDataResample}
\alias{stockDataResample}
\title{Generate random resamples from the passed datasets}
\usage{
stockDataResample(X, N_stock = 50, T_sample = 252 * 2,
  N_dataset = 10, check_monotone = TRUE)
}
\arguments{
\item{X}{a list of xts objects matching the return of function `stockDataDownload()`.}

\item{T_sample}{the length of observation.}

\item{check_monotone}{a logical value indicating whether to check the monotone missing pattern of `X`.}

\item{N_sample}{the number of stocks in each sample.}

\item{N_datasets}{the number of wanted data sets.}
}
\value{
a list of subsets randomly resampled from `X`.
}
\description{
Generate random resamples from the passed datasets
}
\author{
Rui ZHOU & Daniel P. Palomar
}
