% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockDataDownload.R
\name{stockDataResample}
\alias{stockDataResample}
\title{Generate random resamples from financial data}
\usage{
stockDataResample(
  X,
  N_sample = 50,
  T_sample = 2 * 252,
  num_datasets = 10,
  rm_stocks_with_na = TRUE
)
}
\arguments{
\item{X}{List of \code{xts} objects matching the structure returned by the function \code{\link{stockDataDownload}}.}

\item{N_sample}{Desired number of financial instruments in each resample.}

\item{T_sample}{Desired length of each resample (consecutive samples with a random initial time).}

\item{num_datasets}{Number of resampled datasets (chosen randomly among the financial instrument universe).}

\item{rm_stocks_with_na}{Logical value indicating whether to remove instruments with inner missing 
values (ignoring leading and trailing missing values). Default is \code{TRUE}.}
}
\description{
This function is deprecated. Use instead \code{\link{financialDataResample}()}.
}
