% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsNtables.R
\name{summaryBarPlot}
\alias{summaryBarPlot}
\title{Plots a barplot of a backtest summary}
\usage{
summaryBarPlot(res_summary, measures = NULL, type = c("ggplot2",
  "simple"), mar = c(3, 3, 3, 11), inset = c(-0.3, 0),
  legend_loc = "right", ...)
}
\description{
Plots a barplot of a backtest summary
}
\examples{
# After running a backtest and extracting a performance summary, one would get:
bt <- portfolioBacktest(...)
res_summary_median <- backtestSummary(bt, summary_fun = median)

# Here let's generate fake random data for simplicity:
mat <- matrix(runif(4*6), 4, 6)
rownames(mat) <- c("Sharpe ratio", "max drawdown", "annual return", "annual volatility")
colnames(mat) <- c("portfolio 1", "portfolio 2", "portfolio 3", "portfolio 4", "portfolio 5", "portfolio 6")
res_summary_median <- list(performance_summary = mat)

# Now we can obtain the table
summaryBarPlot(res_summary_median, measures = c("max drawdown", "annual volatility"))
summaryBarPlot(res_summary_median, measures = c("max drawdown", "annual volatility"), type = "ggplot2")

}
\author{
Daniel P. Palomar and Rui Zhou
}
