% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsNtables.R
\name{summaryTable}
\alias{summaryTable}
\title{Shows a table of a backtest summary}
\usage{
summaryTable(res_summary, measures = NULL, type = c("simple", "DT",
  "grid.table"))
}
\description{
Shows a table of a backtest summary
}
\examples{
# After running a backtest and extracting a performance summary, one would get:
bt <- portfolioBacktest(...)
res_summary_median <- backtestSummary(bt, summary_fun = median)

# Here let's generate fake random data for simplicity:
mat <- matrix(runif(4*6), 4, 6)
rownames(mat) <- c("Sharpe ratio", "max drawdown", "annual return", "annual volatility")
colnames(mat) <- c("portfolio 1", "portfolio 2", "portfolio 3", "portfolio 4", "portfolio 5", "portfolio 6")
res_summary_median <- list(performance_summary = mat)

# Now we can obtain the table
summaryTable(res_summary_median, measures = c("max drawdown", "annual volatility"))
summaryTable(res_summary_median, measures = c("max drawdown", "annual volatility"), type = "DT")

}
\author{
Daniel P. Palomar and Rui Zhou
}
