% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsNtables.R
\name{summaryTable}
\alias{summaryTable}
\title{Shows a table of a backtest summary}
\usage{
summaryTable(res_summary, measures = NULL, type = c("simple", "DT",
  "grid.table"), order_col = NULL, order_dir = "asc")
}
\description{
Shows a table of a backtest summary
}
\examples{
library(portfolioBacktest)
data("dataset10")  # load dataset

# define your own portfolio function
quintile_portfolio <- function(data) {
  X <- diff(log(data$adjusted))[-1]  
  N <- ncol(X)
  ranking <- sort(colMeans(X), decreasing = TRUE, index.return = TRUE)$ix
  w <- rep(0, N)
  w[ranking[1:round(N/5)]] <- 1/round(N/5)
  return(w)
}

# do backtest
bt <- portfolioBacktest(list("Quintile" = quintile_portfolio), dataset10,
                        benchmark = c("uniform", "index"))

# Now we can obtain the table
res_summary_median <- backtestSummary(bt)
summaryTable(res_summary_median, measures = c("max drawdown", "annual volatility"))
summaryTable(res_summary_median, measures = c("max drawdown", "annual volatility"), type = "DT")

}
\author{
Daniel P. Palomar and Rui Zhou
}
